[Sync] Fix undelete conflict resolution
Previously we were recreating bookmarks when undeleting, and resetting the
version of non-bookmarks. Neither of these are necessary. We now reuse the
current version (so the server knows we are the most up to date versin)
and no longer re-creae bookmarks (because we reuse the existing id, the
write is treated as a modify, not a create).

BUG=154884
TEST=with three signed in clients A, B, C, delete a bookmark on A. On B and C,
before they recieve the notification, modify that same bookmark. Ensure no
duplicates are created.


Review URL: https://chromiumcodereview.appspot.com/11204002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163451 0039d316-1c4b-4281-b951-d872f2087c98