Add new vpn system tray ui. This implements most of the UI features, separate the VPN items out of general network configuration, adding a new uber tray item for vpn, etc. The network icon part for both Network and vpn still needs to be adjusted to meet M24 requirement.
Since we are going to refactor network library and deprecated many of the network API I used, I tried to make a minimum change  to separate the vpn out of the general network settings. Please take a look to see if this will work as a short term workaround.

BUG=129044
TBR=sky

Review URL: https://codereview.chromium.org/11192061

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163676 0039d316-1c4b-4281-b951-d872f2087c98