Linux: don't always launch sandbox renderer tests.
If the sandbox has not been initialized, we should not launch the renderer
sandbox tests. This can happen in only one case: if the Zygote is not used and
the sandbox is de-activated.

We make RunSandboxTests() take an extra no_sandbox parameter to match
InitSandboxTests.

BUG=157470
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11175074

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163951 0039d316-1c4b-4281-b951-d872f2087c98