Lose chrome/renderer/prerender/prerendering_support.*
Move the prerendering_support. functionality into the prerender dispatcher, which is the logical place to handle messages from either the browser or WebKit.

This is being done in preparation for adding events on prerenders, which require more connection between the WebKit/browser messaging.

There's no test, as this should be refactoring only.

BUG=none
R=mmenke@chromium.org
TBR=sky@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11288002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163960 0039d316-1c4b-4281-b951-d872f2087c98