Fix windows GPU driver date related blacklisting.
This is a regression introduced in r164183.

Collected GPU driver dates on Win are in the format of "mm-dd-yyyy", so we need to re-order them into {yyyy, mm, dd} before making version comparison.

The original test only covers one way, so now we expand the test to cover both ways so it won't regress again.

BUG=158092
TEST=content_unittests
Review URL: https://codereview.chromium.org/11308016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@164424 0039d316-1c4b-4281-b951-d872f2087c98