[Android] Use resource throttle instead of HandleExternalProtocol.
Chrome on Android uses a throttle-based mechanism to intercept links
so that the user may choose to run an Android application instead of
loading the link in the browser. This mechanism works for both
internal (e.g. http:, https:) as well as external protocols (e.g.
tel:, geo:).

Instead of using the HandleExternalProtocol path way to handle
external protocols, it is better to have both internal and external
protocols overridden through the same mechanism. This has the
added benefit that an error page can be shown in Chrome on Android
if there is no external handler available for the given protocol, which
is not easily done using the HandleExternalProtocol mechanism.

BUG=156044


Review URL: https://chromiumcodereview.appspot.com/11265013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@164717 0039d316-1c4b-4281-b951-d872f2087c98