The layer animator should hang onto weak, rather than raw, pointers to running animations.
To ensure that these pointers don't go stale, internally, the layer animator
will use weak pointers and check that the instances they point to are still
live before using them.

A consequence of this change is that the animator methods have to be prepared
for NULL sequences (RunningAnimation::sequence() can now return NULL if the
sequence is dead).


BUG=


Review URL: https://chromiumcodereview.appspot.com/11300018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@164799 0039d316-1c4b-4281-b951-d872f2087c98