cras_output: Treat negative latency as 0.
If we are too slow to render, the calculated latency can be negative.
It is then converted to a (big) unsigned number. So the browser thinks
the stream is still active for a long time.

BUG=chromium-os:35393
TEST=ctrl-alt-t to open hterm, open another tab for
chrome://media-internals, press tab in hterm to make it beep, then
check media-internals to see the stream is paused after playing.


Review URL: https://chromiumcodereview.appspot.com/11340004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@165118 0039d316-1c4b-4281-b951-d872f2087c98