Flash Mac: Fix escape to exit fullscreen
On tmz.com, pressing escape to exit fullscreen Flash would cause the currently playing movie to close.

The problem was that the fullscreen window was closing on a key down event and the key up event was going to the parent site.

My fix is to suppress the key up event on the parent view.

BUG=155492


Review URL: https://chromiumcodereview.appspot.com/11188027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@165615 0039d316-1c4b-4281-b951-d872f2087c98