Revert 165809 - [Autofill] Disable the browser process popup UI for non-Android platforms on M24.
This UI is not baked enough on Windows or Linux to be enabled in M24.  However, it is the only Autofill UI for Android, and should remain on for that platform.

This CL is meant to be committed, merged to M24, then reverted on trunk.

BUG=158371

Review URL: https://codereview.chromium.org/11364058

TBR=isherman@chromium.org
Review URL: https://codereview.chromium.org/11367079

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@165811 0039d316-1c4b-4281-b951-d872f2087c98