Android: disable use of libyuv.
The Anrdoid WebView build doesn't have libyuv, so the code depending on
it was broken there. It doesn't look like we need this on Android at
all, so just disable it based on OS for now; this can be revisited if
it's needed later.

BUG=

Review URL: https://codereview.chromium.org/11366084

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@166190 0039d316-1c4b-4281-b951-d872f2087c98