LayerImpl::tryScroll must convert its content-space point to layer space before comparing to the nonFastScrollableRegion.
The hit test resulting point that the tryScroll() method gets is in content
space, since it is the result of mapping via the inverse screenSpaceTransform.
However, the nonFastScrollableRegion is in layer space, so it should not compare
the point against the region directly. We add a conversion from content to layer
space before doing the test.

This is now covered by the following tests. Because the layer is given a
contentsScale, the tests fail without the above change made to tryScroll.

Tests:
cc_unittests:LayerTreeHostImplTest.nonFastScrollableRegionBasic
cc_unittests:LayerTreeHostImplTest.nonFastScrollableRegionWithOffset

R=enne
BUG=159676


Review URL: https://chromiumcodereview.appspot.com/11377006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@166353 0039d316-1c4b-4281-b951-d872f2087c98