First couple of steps of improving the extension/app update process.
For Apps and Extensions with a lazy background page, this will delay installing
an update until the background page gets unloaded (or next chrome restart).
Also adds a chrome.runtime.onUpdateAvailable event which is fired when an update is
available but isn't being installed immediately because the app is not idle.
It doesn't change anything yet for Extensions with a persistent background page
(or extensions without a background page at all). Those still get updated as soon
as an update is found.

BUG=37971
BUG=143260
TBR=nick@chromium.org 

Review URL: https://chromiumcodereview.appspot.com/11293002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@166358 0039d316-1c4b-4281-b951-d872f2087c98