retry r166570: now with clang compile fix.
[views] continue hacking away at autofill dialog (#4).

in this CL, accept inputs (which fields a page asks for) and collect outputs (values for those fields).

The input data structure is a vector of AutofillFieldType (order is not important). The output structure is a mapping from these field types to string values supplied by the user (or filled in by suggestion). This input and output format will need to be supplied and accepted by the autofill agent.

BUG=157270, 157273

Review URL: https://codereview.chromium.org/11363135

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@166587 0039d316-1c4b-4281-b951-d872f2087c98