Implement hit test related methods in Android WebView
The public Android WebView APIs are WebView.getHitTestResult,
WebView.requestFocusNodeHref, and WebView.requestImageRef.

These APIs are mostly used for creating context menus but have
become broken APIs as the Android evolved from the single
threaded initial version of WebView.

* The names are misleading since the value retrieved is based both on
  a combination of touch events and FocusedNodeChanged callback
  from WebKit.
* The methods are synchronous and there are inherently racy. This is
  the case with with the latest version of Android WebView as well.
  However this may become more of a problem as Chromium is
  multi-processed.
* The methods are inefficient since work needs to be done even if the
  client application never call these methods.

This is the first part of largely replicating the broken code with tests
and extensive comments. The goal is to replicate existing logic but
there are probably corner cases that are not matched.

Overall flow is on a touch event or focused node changed event, perform
a WebKit hit test, get the relevant result data and save it in the browser.
The return what is saved when apps query for the data.

Work still needs to be done after this:
* Implement updating hit test data after tab-ing to a link
* Hook up content detection code in content/
* Implement focus highlighting with DPAD controls.

BUG=

All bots are green except unrelated instrumentation tests.
AndroidWebView instrumentation tests are green.
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11360037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@166712 0039d316-1c4b-4281-b951-d872f2087c98