This is the first pass at making GetIPConfigs asynchronous.
This renames the old GetIPConfigs to GetIPConfigsAndBlock, and deprecates it.

Several sites which used to use the old blocking GetIPConfigs have been rewritten to use the non-blocking asynchronous version.

There are still two places that were too complex to convert:
MoreMenuModel::InitMenuItems
SystemTrayDelegate::GetNetworkAddresses

Both of those sites rely on getting their results immediately and
would need their callers to change significantly to switch to async.

In order to allow the IPConfigs to be fetched asynchronously, we need to both get the shill properties for the service and the ipconfigs, so we have to have two layers of async calls.  This requires that we convert DictionaryValueCallbacks to use scoped_ptrs so we can avoid a deep copy of the properties.

BUG=chromium:147657
TEST=ran on device, tested network menus and settings.  Ran unit tests.


Review URL: https://chromiumcodereview.appspot.com/11367048

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@167427 0039d316-1c4b-4281-b951-d872f2087c98