Add a ShadowingAtExitManager to ExtensionServiceTestSimple.Enabledness.
This is required to correctly tear down the PluginService instance created
during the test.  The test creates a mock PluginList on the stack and passes
it to the PluginService singleton, so if the singleton isn't torn down then
later tests will crash or hang accessing a pointer into the stack that is no
longer valid.

BUG=159754,149157,155541,160592


Review URL: https://chromiumcodereview.appspot.com/11312204

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@167529 0039d316-1c4b-4281-b951-d872f2087c98