Handle NinePatchLayer Resource Rebuilding.
When we minimize the application our PrioritizedResourceManager dies.  This clears the
PrioritizedTexture state and decouples the texture from the manager.  When we bring our
application back to the foreground a new PrioritizedResourceManager is created, but we
still hold onto our old PrioritizedResource.  Trying to reupload a bitmap on this
resource fails due to a variety of DCHECK/CHECKs at this point.  Now if the
PrioritizedResource has no PrioritizedResourceManager we rebuild the resource.

BUG=http://crbug.com/161278


Review URL: https://chromiumcodereview.appspot.com/11420021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@168122 0039d316-1c4b-4281-b951-d872f2087c98