Making it easier to hook a ContentViewCore to the SurfaceView.
This CL changes the way we deal with WebLayers, so that a ContentViewCore does
not need native code anymore to hook with the surface view.
The ContentViewCore now keeps track of the WebLayers as they are
attached/detached. The ContentViewRenderView is used to render a specific
ContentView by retrieving the ContentViewCore WebLlayer and adding it to the
compositor.

BUG=
TEST=ContentShell and ChromiumTestShell should still work.


Review URL: https://chromiumcodereview.appspot.com/11293264

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@168166 0039d316-1c4b-4281-b951-d872f2087c98