Removing environment variable as default value for 'cros_remote'.
This causes my benchmarks to crash when running
./run_multipage_benchmarks --browser=release scrolling_benchmark page_sets/top_25.json,
for example, as long as my REMOTE environment variable is set. Other approaches
to fix this could be to remove the check in browser_finder.py
(if options.browser_type != 'cros-chrome' and options.cros_remote != None)
or to add a separate check to only fail if this was specified by command-line
and not just by environment variable.

Personally, I think we can just do without this default value, but let me know
what you think.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11417013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@168200 0039d316-1c4b-4281-b951-d872f2087c98