Refactor the URLResponseInfo to use new design
This puts all of the URLResponseInfo attributes in a struct so it can be sent over IPC in one message rather than requiring one sync IPC per attribute access.

This includes a new example of streaming to a file that I used to do some manual tests of this change.

I added and improved some documentation that I noticed when trying to write the example.

Review URL: https://codereview.chromium.org/10993031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@168273 0039d316-1c4b-4281-b951-d872f2087c98