chromeos: Make --load-opencryptoki unnecessary on devices.
We unconditionally pass --load-opencryptoki to Chrome on
Chrome OS devices.  This change makes it not be required in
those cases, but still leaves the flag so it can be forced
when running elsewhere.  (I can't unconditionally turn it
on; on my workstation, it makes us repeatedly fail and often
triggers the DCHECK(request_task_.is_null()) in
CertLibraryImpl's d'tor.)

BUG=143164


Review URL: https://chromiumcodereview.appspot.com/11420065

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@168641 0039d316-1c4b-4281-b951-d872f2087c98