Stop using shell::GetInstance()->system_tray() in system tray items
This CL provides all SystemTrayItem objects with a pointer to the parent
SystemTray and all Tray*View objects with a pointer to the parent
SystemTrayItem. This allows the objects to walk up the chain and access
the correct system tray once multiple system trays are introduced.

UpdateNagger::~UpdateNagger() in tray_update.cc continues to go through
shell::GetInstance() as the refactoring needed to support multiple system
trays here is more complex and out of my expertise.

BUG=152928

Review URL: https://chromiumcodereview.appspot.com/11415014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@169141 0039d316-1c4b-4281-b951-d872f2087c98