SECCOMP-BPF: Added support for greylisting of system calls.
In addition to a Sandbox::Trap() handler, we now have a Sandbox::UnsafeTrap()
handler. This feature should only be used for debugging purposes as it subverts
the security of the sandbox. But it is useful to track down problems with the
sandboxing policy. Within an unsafe trap handler, all sandbox restrictions are
lifted. This, for example, allows us to allow system calls that would normally
be denied by the policy, but to log their arguments, return value, and call stack.

N.B.: this is the second attempt at submitting this CL. See https://chromiumcodereview.appspot.com/11363212/
      for previous code reviews


BUG=130662
TEST=sandbox_linux_unittests
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/11419121

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@169213 0039d316-1c4b-4281-b951-d872f2087c98