[Sync] Refactor control types initialization
We now have all the control type initialization happen on the sync 
thread, avoiding the ConfigureDataTypes method which would stuff 
the nigori node into any download. This allows us to ensure only those 
control types without progress markers are downloaded as part of backend 
initialization. 

To accomplish this, HandleInitializationCompletedOnFrontendLoop no longer 
operates as a state machine. Instead, we only call it when initialization has 
completely finished (either as a success or a failure). 

Note that theoretically HandleSyncManagerInitializationOnFrontendLoop could 
also be removed, in favor of having all initialization happen on the sync thread, 
but it's useful to keep it around for the purpose of doing pre-initialization 
completion work in tests. 

BUG=161824


Review URL: https://chromiumcodereview.appspot.com/11419171

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@169773 0039d316-1c4b-4281-b951-d872f2087c98