Remove redundant call to DevToolsWindow::UpdateFrontendDockSide()
This redundant call in Show will cause errors like "Uncaught ReferenceError:
InspectorFrontendAPI is not defined" in the case DevTools has not
loaded yet. In the case that it is loaded, DoAction will call the
function, and when it isn't loaded the observer will call DoAction.

BUG=162799
R=pfeldman@chromium.org
TEST=covered by browser_tests, e.g. ChromeRenderProcessHostTest.DevToolsOnSelfInOwnProcess


Review URL: https://chromiumcodereview.appspot.com/11280180

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@169941 0039d316-1c4b-4281-b951-d872f2087c98