Desktop aura: Break aura::Window::SetParent in two.
Previously, SetParent will query a global StackingClient object if NULL is
passed to it. Since we want different behavior on the desktop and on ash, we
need to break that. It has been replaced with two methods:

- SetParentTo(), which takes an aura window and does the parenting.
- SetDefaultParentByTargetRoot(), which takes a RootWindow as context and
asks the StackingClient on said RootWindow where it should parent the
window.

The problem is that people have relied for a long time on what amounts to a
global variable. This is the first of several patches that pass around a
context RootWindow. This patch focuses on ash/ unittests mostly. Later patches
will focus on threading context through Widget creation. The last patch in this
series will pull out the global stacking client interface and add DCHECKs so
NULL can't be passed to either of these methods on Window.

BUG=161882


Review URL: https://chromiumcodereview.appspot.com/11421006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@170049 0039d316-1c4b-4281-b951-d872f2087c98