net: add special case errors for certificates.
Over thanksgiving I tried adding a public key used by some malware to the
CRLSet. However, the malware appears to be fairly wide spread and a number of
users, reasonably, were somewhat confused by the message.

In the future it seems that we probably want to be able to set a URL and
message in the CRLSet so that we can provide better direction.

This change adds the infrasture for that but, temporarily, hardcodes the
specific malware case. Unfortunately, interstitial pages cannot navigate, open
new windows nor load iframes without lots of work in content/. Therefore the
URL (which is intended to provided a translated message for the user) has to be
copied and pasted. Hopefully can be pursude one of the content folks to improve
this for the future.

BUG=none

Review URL: https://codereview.chromium.org/11413169

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@170083 0039d316-1c4b-4281-b951-d872f2087c98